+++ src/helpers/Url/Url.js
@@ -94,14 +94,17 @@ export function getView(url) {
  */
 export function flattenToAppURL(url) {
   const { settings } = config;
+  const isPrefixPath = url?.indexOf(settings.prefixPath) === 0;
+
   return (
     url &&
-    url
+    `${isPrefixPath ? '' : settings.prefixPath}${url
       .replace(settings.internalApiPath, '')
       .replace(settings.apiPath, '')
-      .replace(settings.publicURL, '')
+      .replace(settings.publicURL, '')}`
   );
 }
+
 /**
  * Given a URL it remove the querystring from the URL.
  * @method stripQuerystring
@@ -154,11 +157,12 @@ export const isCmsUi = memoize((currentPathname) => {
  */
 export function flattenHTMLToAppURL(html) {
   const { settings } = config;
+  const replacer = config.settings.prefixPath ?? '';
   return settings.internalApiPath
     ? html
-        .replace(new RegExp(settings.internalApiPath, 'g'), '')
-        .replace(new RegExp(settings.apiPath, 'g'), '')
-    : html.replace(new RegExp(settings.apiPath, 'g'), '');
+        .replace(new RegExp(settings.internalApiPath, 'g'), replacer)
+        .replace(new RegExp(settings.apiPath, 'g'), replacer)
+    : html.replace(new RegExp(settings.apiPath, 'g'), replacer);
 }
 
 /**