+++ src/store.js
@@ -12,12 +12,19 @@ import {
   blacklistRoutes,
   protectLoadStart,
   protectLoadEnd,
+  crashReporter,
+  prefixPathRoot,
   loadProtector,
 } from '@plone/volto/middleware';
 
 const configureStore = (initialState, history, apiHelper) => {
+  const unlisten = history.listen((location, action) => {
+    // location is an object like window.location
+    console.log(action, location.pathname, location.state);
+  });
   let stack = [
     blacklistRoutes,
+    prefixPathRoot(history),
     protectLoadStart,
     routerMiddleware(history),
     thunk,